g4-bit Windows build targets for the validator libraries
Review URL: http://codereview.chromium.org/557005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1419 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 00bef40b98177c9a7075261c012aa36073869c40

+++ src/trusted/validator_x86/nc_read_segment.c
@@ -51,7 +51,7 @@
  */
 #define MAX_INPUT_LINE 4096
 
-static void ConvertHexToByte(char mini_buf[3], int mini_buf_index,
+static void ConvertHexToByte(char mini_buf[3], size_t mini_buf_index,
                              uint8_t* mbase, size_t mbase_size,
                              size_t* count) {
   mini_buf[mini_buf_index] = '\0';
